update seed script
