+++ db/__init__.py
@@ -1,3 +1,2 @@
 from .init_db import *
 from .db_utils import *
-from .seed_db import *
